refactor: admin-handler.ts OOP 拆分 + Chat 独立端点
将 953 行的 God Class admin-handler.ts 按域拆分为独立 Handler：
- WsDomainHandler 抽象基类 + HandlerContext 依赖注入
- ConfigHandler / StatusHandler / PluginHandler / LogHandler / SessionHandler
- ChatWsHandler 独立为 /ws/chat 端点
- ConfigStore 提取配置读写逻辑
- AdminWsHandler 瘦身为 229 行的薄 Router

Closes #56
