+++ apps/server/src/gateway/ws/handlers/index.ts
@@ -1,0 +1,10 @@
+/**
+ * Domain Handlers 导出
+ */
+
+export { WsDomainHandler } from './base.js'
+export { ConfigHandler } from './config-handler.js'
+export { StatusHandler } from './status-handler.js'
+export { PluginHandler } from './plugin-handler.js'
+export { LogHandler } from './log-handler.js'
+export { SessionHandler } from './session-handler.js'
