+++ openspec/changes/archive/2026-03-31-refactor-admin-handler-oop/.openspec.yaml
@@ -1,0 +1,2 @@
+schema: spec-driven
+created: 2026-03-31
